refactor(button): move component into src directory
