+++ packages/button/index.ts
@@ -1,9 +1,9 @@
-export { default as Button } from "./Button/Button"
-export { default as IconButton } from "./Button/IconButton"
-export { default as DirectionalLink } from "./Button/DirectionalLink"
-export { default as PaginationLink } from "./Button/PaginationLink"
+export { default as Button } from "./src/Button/Button"
+export { default as IconButton } from "./src/Button/IconButton"
+export { default as DirectionalLink } from "./src/Button/DirectionalLink"
+export { default as PaginationLink } from "./src/Button/PaginationLink"
 export type {
   ButtonRef,
   CustomButtonProps,
-} from "./Button/components/GenericButton"
-export type { ButtonProps } from "./Button/Button"
+} from "./src/Button/components/GenericButton"
+export type { ButtonProps } from "./src/Button/Button"
